refactor!: split cli
