+++ .github/workflows/build-test.yml
@@ -7,14 +7,7 @@ on:
     branches-ignore:
       - 'main'
     paths:
-      - 'src/bsm_api_client/**'
-      - 'tests/**'
-      - 'pyproject.toml'
-  #pull_request:
-  #  paths:
-  #    - 'src/bsm_api_client/**'
-  #    - 'tests/**'
-  #    - 'pyproject.toml'
+      - 'packages/**'
 
 jobs:
   build-and-test:
@@ -41,21 +34,24 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install build
-        pip install --upgrade -e .[dev]
+        pip install --upgrade -e ./packages/bsm-api-client[dev]
+        pip install --upgrade -e ./packages/bsm-cli[dev]
 
     - name: Run Black
-      run: black --check src/bsm_api_client tests
+      run: black --check packages
 
     - name: Run Isort
-      run: isort --check-only src/bsm_api_client tests
+      run: isort --check-only packages
 
     - name: Run Mypy
       continue-on-error: true
-      run: mypy src/bsm_api_client tests
+      run: mypy packages
 
     - name: Run Flake8
       continue-on-error: true
-      run: flake8 src/bsm_api_client tests
+      run: flake8 packages
 
     - name: Run Pytest
-      run: python -m pytest
+      run: |
+        python -m pytest packages/bsm-api-client
+        python -m pytest packages/bsm-cli
