+++ packages/bsm-api-client/pyproject.toml
@@ -1,4 +1,3 @@
-# pyproject.toml
 [build-system]
 requires = ["setuptools>=80.9", "setuptools_scm"]
 build-backend = "setuptools.build_meta"
@@ -40,24 +39,18 @@ dev = [
     "mypy >=1.10.0,<2.2",
     "bedrock-server-manager >=3.10.0b2,<3.11.0",
 ]
-cli = [
-    "click >=8.2.0,<8.5",
-    "questionary >=2.1.0,<2.2",
-]
 
 [project.urls]
 "Homepage" = "https://github.com/DMedina559/bsm-api-client"
 "Bug Tracker" = "https://github.com/DMedina559/bsm-api-client/issues"
-"Changelog" = "https://github.com/DMedina559/bsm-api-client/blob/main/docs/CHANGELOG.md"
-
-[project.scripts]
-bsm-api-client = "bsm_api_client.cli.__main__:cli"
+"Changelog" = "https://github.com/DMedina559/bsm-api-client/releases"
 
 [tool.setuptools.packages.find]
 where = ["src"]
 include = ["bsm_api_client*"]
 
 [tool.setuptools_scm]
+root = "../.."
 local_scheme = "no-local-version"
 
 [tool.black]
