+++ packages/bsm-api-client/tests/integration/test_manager_and_server_info.py
