+++ packages/bsm-cli/src/bsm_cli/addon.py
@@ -2,6 +2,7 @@ import os
 
 import click
 import questionary
+from bsm_cli.decorators import monitor_task, pass_async_context
 from questionary import Separator
 
 from bsm_api_client.models import (
@@ -10,8 +11,6 @@ from bsm_api_client.models import (
     AddonSubpackPayload,
     FileNamePayload,
 )
-
-from .decorators import monitor_task, pass_async_context
 
 
 @click.group()
