+++ packages/bsm-cli/src/bsm_cli/content.py
@@ -1,9 +1,8 @@
-# src/bsm_api_client/cli/content.py
+# src/bsm_cli/content.py
 """CLI commands for content management."""
 
 import click
-
-from .decorators import pass_async_context
+from bsm_cli.decorators import pass_async_context
 
 
 @click.group()
