+++ packages/bsm-cli/src/bsm_cli/properties.py
