+++ packages/bsm-cli/src/bsm_cli/server.py
@@ -3,14 +3,13 @@ import os
 
 import click
 import questionary
+from bsm_cli.allowlist import interactive_allowlist_workflow
+from bsm_cli.decorators import monitor_task, pass_async_context
+from bsm_cli.permissions import interactive_permissions_workflow
+from bsm_cli.properties import interactive_properties_workflow
 
 from bsm_api_client.exceptions import AuthError
 from bsm_api_client.models import CommandPayload, InstallServerPayload
-
-from .allowlist import interactive_allowlist_workflow
-from .decorators import monitor_task, pass_async_context
-from .permissions import interactive_permissions_workflow
-from .properties import interactive_properties_workflow
 
 
 def _print_server_table(servers):
