+++ packages/bsm-cli/src/bsm_cli/users.py
@@ -1,7 +1,6 @@
 import click
 import questionary
-
-from .decorators import pass_async_context
+from bsm_cli.decorators import pass_async_context
 
 
 async def interactive_user_workflow(ctx, client):  # noqa: C901
