+++ packages/bsm-cli/src/bsm_cli/world.py
@@ -2,10 +2,9 @@ import os
 
 import click
 import questionary
+from bsm_cli.decorators import monitor_task, pass_async_context
 
 from bsm_api_client.models import FileNamePayload
-
-from .decorators import monitor_task, pass_async_context
 
 
 @click.group()
