+++ packages/bsm-cli/tests/test_cli_websocket.py
@@ -2,9 +2,9 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import click
 import pytest
+from bsm_cli.decorators import monitor_task
+from bsm_cli.server import list_servers
 
-from bsm_api_client.cli.decorators import monitor_task
-from bsm_api_client.cli.server import list_servers
 from bsm_api_client.websocket_client import WebSocketClient
